init structure
